flask/requirements.txt
psycopg2-binary==2.9.3
Flask==2.0.1
scikit-surprise==1.1.3
Werkzeug==2.0.1
pandas
numpy